requirements.txt
PyMuPDF
pdf2image
Pillow
tqdm
google-cloud-vision
requests
google-generativeai